Overhaul the API
The plugin system subsumed the functionality of building html.nix files,
so now html.nix files are just another "plugin" (that term is no longer
used). What was the attrset of plugins before is currently called a
"build specification", or "spec" for short.

There is a library for dealing with paths, which is used in the project
and exposed to the user.